% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_annotate.R
\name{berlin_annotate}
\alias{berlin_annotate}
\title{Perform SingleR annotation of Seurat object}
\usage{
berlin_annotate(object = NULL, verbose = TRUE)
}
\arguments{
\item{object}{Seurat object. If NULL, must supply count data.}

\item{verbose}{Boolean. To show progress, TRUE, else FALSE.}
}
\value{
A Seurat object.
}
\description{
Perform SingleR annotation of Seurat object
}
