% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_cluster.R
\name{berlin_cluster}
\alias{berlin_cluster}
\title{Perform dimension reduction and clustering analysis on Seurat object}
\usage{
berlin_cluster(
  object = NULL,
  assay = "RNA",
  seed = 42,
  verbose = TRUE,
  resolution = c(0.5, 1, 1.5, 2),
  pca_npcs = 30,
  tsne_dims = 1:30,
  umap_dims = 1:10,
  neighbor_dims = 1:30
)
}
\arguments{
\item{object}{Seurat object. If NULL, must supply count data.}

\item{assay}{String. Name of the initial/default assay.}

\item{seed}{Numeric. Random seed, default is 42. Setting to NULL will remove seed.}

\item{verbose}{Boolean. To show progress, TRUE, else FALSE.}

\item{resolution}{Numeric vector. Above one to obtain a larger number of communities, and below one to obtain a smaller number.}

\item{pca_npcs}{Numeric. Total number of principal components to compute for PCA plot. Default is 30.}

\item{tsne_dims}{Numeric sequence. Dimensions to use as input features for TSNE plot. Default 1:30.}

\item{umap_dims}{Numeric sequence. Dimensions to use as input features for UMAP. Default 1:10.}

\item{neighbor_dims}{Numeric sequence. Dimension of reduction to use as input to find neighbors.}

\item{doublet_pN}{Numeric. The number of generated artificial doublets, expressed as a proportion of the merged real-artificial data. Default is set to 0.25.}

\item{doublet_pK}{Numeric. The principal component neighborhood size to compute pANN, expressed as a proportion of the merged real-artificial data. Default is 0.09.}

\item{doublet_prop}{Numeric. Estimated proportion of homotypic doublets. Default is 0.04.}

\item{doublet_PCs}{Numeric. Number of statistically-significant principal components.}
}
\value{
A Seurat object.
}
\description{
Perform dimension reduction and clustering analysis on Seurat object
}
