% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_get_started.R
\name{berlin_filter}
\alias{berlin_filter}
\title{Filter Seurat object by percent mitochondria or number of features}
\usage{
berlin_filter(
  seurat = NULL,
  species = "human",
  percent_mt = NULL,
  minFeatures = 500
)
}
\arguments{
\item{seurat}{Seurat object.}

\item{species}{String denoting if the data is from "human" or "mouse".}

\item{percent_mt}{Numeric. Percent mitochondria to filter out. Default is NULL.}

\item{minFeatures}{Numeric. Minimum number of features a barcode/cell must have. Default to 500.}
}
\value{
Seurat object of filtered features.
}
\description{
Filter Seurat object by percent mitochondria or number of features
}
