% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_marker_analysis.R
\name{berlin_jaccard}
\alias{berlin_jaccard}
\title{Assign Cluster to Geneset Using Jaccard Formula}
\usage{
berlin_jaccard(
  object = NULL,
  markers_res = NULL,
  geneset = NULL,
  cluster_col = "seurat_clusters",
  res_col = NULL,
  assay = "RNA",
  min.pct = 0.25,
  adjpval_cutoff = 0.05,
  log2fc_cutoff = 1,
  verbose = TRUE,
  return_seurat = FALSE
)
}
\arguments{
\item{object}{Seurat object. If NULL, must supply count data.}

\item{markers_res}{Data.frame. Result data from Seurat's FindAllMarkers() function.}

\item{geneset}{Data.frame or list. Two column data frame with first column consisting of geneset names and second column of the genes withing that geneset. One gene per row, the geneset name can repeat in the first column. If list input, this would be a named list where the names are the geneset names and each element a vector of genes as strings.}

\item{cluster_col}{String. Column name from meta data of primary identified clusters. Default to 'seurat_clusters'.}

\item{res_col}{String. Desired column name for jaccard results. Defaults to the name of the cluster column appeneded with '_Jaccard_Classification'.}

\item{assay}{String. Name of the initial/default assay. Default is RNA.}

\item{min.pct}{Numeric. Only test genes that are detected in a minimum fraction of cells. Default is 0.25.}

\item{adjpval_cutoff}{Numeric. An adjusted p-value value cutoff for identifying significantly expressed genes. Default is 0.05.}

\item{log2fc_cutoff}{Numeric. An absolute value of a log2 fold change cutoff to identify up and down regulated genes. Default is 0.}

\item{verbose}{Boolean. To show progress, TRUE, else FALSE.}

\item{return_seurat}{Boolean. If TRUE, a Seurat object will be the output with the results within the stored in the Misc slot.}
}
\value{
a Data.frame or seurat object
}
\description{
Assign Cluster to Geneset Using Jaccard Formula
}
