% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_qc.R
\name{berlin_qc}
\alias{berlin_qc}
\title{Perform quality control and initial formatting of Seurat object}
\usage{
berlin_qc(
  object = NULL,
  counts = NULL,
  meta = NULL,
  assay = "RNA",
  project_name = "BERLIN_Project",
  species = "human",
  percent_mt = NULL,
  minFeatures = 500,
  varFeatures = 2000,
  verbose = TRUE,
  logScale = 10000,
  remove_duplicates = TRUE
)
}
\arguments{
\item{object}{Seurat object. If NULL, must supply count data.}

\item{counts}{A matrix, data.frame, or dgCMatrix. The rows are unique features. The feature name can be the rowname or the first column. Each column is a uniquely labeled barcode/cell.}

\item{meta}{A data.frame. The rows are barcode/cell names, and the column are any additionally supplied metadata. Row names in the metadata need to match the column names of the counts matrix.}

\item{assay}{String. Name of the initial/default assay.}

\item{project_name}{String. Desired project name.}

\item{species}{String. Denoting if the data is from "human" or "mouse".}

\item{percent_mt}{Numeric. Percent mitochondria to filter out. Default is NULL.}

\item{minFeatures}{Numeric. Minimum number of features a barcode/cell must have. Default to 500.}

\item{varFeatures}{Numeric. Number of features to select as top variable features. Default to 2000.}

\item{verbose}{Boolean. To show progress, TRUE, else FALSE.}

\item{logScale}{Numeric. Set the scale to log normalize data. Defualt to 10000.}

\item{remove_duplicates}{Boolean. To remove duplicate features that may be found in the data, use TRUE, else FALSE and the function will notify you if duplicate features are found.}
}
\value{
Seurat object.
}
\description{
Perform quality control and initial formatting of Seurat object
}
