% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_scgate.R
\name{berlin_run_scgate}
\alias{berlin_run_scgate}
\title{Performs scGate function with a single model on a Seurat object}
\usage{
berlin_run_scgate(
  object = NULL,
  model = NULL,
  model_name = NULL,
  assay = "RNA",
  pos.thr = 0.2,
  neg.thr = 0.2,
  ncores = 1,
  seed = 42,
  verbose = TRUE
)
}
\arguments{
\item{object}{Seurat object. If NULL, must supply count data.}

\item{model}{A single scGate model, or a list of scGate models. Please see details for formatting.}

\item{model_name}{String. A name for the input model.}

\item{assay}{String. Name of the initial/default assay. Default to RNA.}

\item{pos.thr}{Numeric. Minimum UCell score value for positive signatures.}

\item{neg.thr}{Numeric. Maximum UCell score value for negative signatures.}

\item{ncores}{Numeric. Number of processors for parallel processing. Default to 1.}

\item{seed}{Numeric. Random seed, default is 42. Setting to NULL will remove seed.}

\item{verbose}{Boolean. To show progress, TRUE, else FALSE.}
}
\value{
A data.frame.
}
\description{
Performs scGate function with a single model on a Seurat object
}
