% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_scgate.R
\name{berlin_scgate}
\alias{berlin_scgate}
\title{Performs ScGate function on a Seurat object with one or multiple models}
\usage{
berlin_scgate(
  object = NULL,
  assay = "RNA",
  model = NULL,
  model_name = NULL,
  pos.thr = 0.2,
  neg.thr = 0.2,
  species = "human",
  ncores = 1,
  seed = 42,
  verbose = TRUE
)
}
\arguments{
\item{object}{Seurat object. If NULL, must supply count data.}

\item{assay}{String. Name of the initial/default assay. Default to RNA.}

\item{model}{List object or data.frame. A single scGate model, or a list of scGate models. Please see \link[scGate]{scGate} for further formatting parameters.}

\item{model_name}{String. A desired name for the input model. Required if only running a single model.}

\item{species}{String. Denoting if the data is from "human" or "mouse". Required if using package provided models.}

\item{ncores}{Numeric. Number of processors for parallel processing. Default to 1.}

\item{seed}{Numeric. Random seed, default is 42. Setting to NULL will remove seed.}

\item{verbose}{Boolean. To show progress, TRUE, else FALSE.}
}
\value{
Seurat object.
}
\description{
Performs ScGate function on a Seurat object with one or multiple models
}
