% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_scgate.R
\name{berlin_scgate_model}
\alias{berlin_scgate_model}
\title{Generate scGate model from parameter data}
\usage{
berlin_scgate_model(params = NULL)
}
\arguments{
\item{params}{A data.frame. The scGate model will be generated based on these parameters. More information about the format in details.}
}
\value{
A R List Object.
}
\description{
Generate scGate model from parameter data
}
\details{
The first column should indicate the cell type label. The following columns should be named to indicate the level and positive or negative indication of the signature.
The level should be a numeric number starting at 1 and the positive and negative indication will be determined by grepping the column name for 'pos' or 'neg'.
For example the column name 'Level2Pos' indicated a column of level 2 positive indication signatures. The user does not need to include the word 'level' and
can write out the word of 'positive' or 'negative' and put the level number after the indicator if preferred.
}
