% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_scgate.R
\name{berlin_scgate_summarize}
\alias{berlin_scgate_summarize}
\title{Summarize scGate annotation across clusters}
\usage{
berlin_scgate_summarize(
  object = NULL,
  celltype_col = NULL,
  cluster_col = "seurat_clusters",
  show_unassigned = TRUE,
  return_seurat = FALSE
)
}
\arguments{
\item{object}{Seurat object.}

\item{celltype_col}{String. Name of column containing predicted cell type names. If not supplied it will be predicted based on string matching.}

\item{cluster_col}{String. Name of column containing cluster information.}

\item{show_unassigned}{Boolean. If TRUE, unassigned or NA cell types will be included.}

\item{return_seurat}{Boolean. If TRUE, a Seurat object will be the output with the results within the stored in the Misc slot.}
}
\value{
A data.frame object.
}
\description{
Summarize scGate annotation across clusters
}
