% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_sctype.R
\name{berlin_sctype}
\alias{berlin_sctype}
\title{Generate ScType encrichment score matrix from Seurat object}
\usage{
berlin_sctype(
  object = NULL,
  scaled = TRUE,
  geneset = "immune",
  pos_geneset = NULL,
  neg_geneset = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{Seurat object. If NULL, must supply count data.}

\item{scaled}{Boolean. TRUE if the data has already been scaled. If the data still needs to be scaled, FALSE and the function will scale the data.}

\item{geneset}{String. Input a tissue geneset name from the provided in the package (see details). Run ScType_genesets() function to view available genesets.}

\item{pos_geneset}{List object. A list of positive marker genesets. Each element in the list should be a vector of genes and the name is the geneset name.}

\item{verbose}{Boolean. To show progress, TRUE, else FALSE.}

\item{neg_genesetList}{object. A list of negative marker genesets. Each element in the list should be a vector of genes and the name is the geneset name.}
}
\value{
Matrix
}
\description{
Generate ScType encrichment score matrix from Seurat object
}
