% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_sctype.R
\name{berlin_sctype_summarize}
\alias{berlin_sctype_summarize}
\title{Summarize scType cell annotation across clusters}
\usage{
berlin_sctype_summarize(
  object = NULL,
  scType_cell_col = NULL,
  cluster_col = "seurat_clusters",
  show_unassigned = TRUE,
  return_seurat = FALSE
)
}
\arguments{
\item{object}{Seurat object.}

\item{scType_cell_col}{String. Name of column containing predicted scType cell levele classification. If not supplied it will be predicted based on string matching.}

\item{cluster_col}{String. Name of column containing cluster information. Default to "seurat_clusters".}

\item{show_unassigned}{Boolean. If TRUE, unassigned or NA cell types will be included.}

\item{return_seurat}{Boolean. If TRUE, a Seurat object will be the output with the results within the stored in the Misc slot.}
}
\value{
A data.frame object.
}
\description{
Summarize scType cell annotation across clusters
}
