% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_unsupervised_analysis.R
\name{berlin_unsupervised_analysis}
\alias{berlin_unsupervised_analysis}
\title{Perform dimension reduction and clustering analysis on Seurat object}
\usage{
berlin_unsupervised_analysis(
  object = NULL,
  assay = "RNA",
  seed = 42,
  verbose = TRUE,
  resolution = c(0.5, 1, 1.5, 2),
  run_pca = TRUE,
  run_umap = TRUE,
  run_TSNE = FALSE,
  pca_npcs = 30,
  tsne_dims = 1:30,
  umap_dims = 1:10,
  neighbor_dims = 1:30
)
}
\arguments{
\item{object}{Seurat object. If NULL, must supply count data.}

\item{assay}{String. Name of the initial/default assay.}

\item{seed}{Numeric. Random seed, default is 42. Setting to NULL will remove seed.}

\item{verbose}{Boolean. To show progress, TRUE, else FALSE.}

\item{resolution}{Numeric vector. Above one to obtain a larger number of communities, and below one to obtain a smaller number.}

\item{pca_npcs}{Numeric. Total number of principal components to compute for PCA plot. Default is 30.}

\item{tsne_dims}{Numeric sequence. Dimensions to use as input features for TSNE plot. Default 1:30.}

\item{umap_dims}{Numeric sequence. Dimensions to use as input features for UMAP. Default 1:10.}

\item{neighbor_dims}{Numeric sequence. Dimension of reduction to use as input to find neighbors.}
}
\value{
A Seurat object.
}
\description{
Perform dimension reduction and clustering analysis on Seurat object
}
