% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_qc.R
\name{date_to_gene}
\alias{date_to_gene}
\title{Convert excel dates to gene symbols}
\usage{
date_to_gene(vec, mouse = FALSE)
}
\arguments{
\item{vec}{Character vector of gene symbols.}

\item{mouse}{Boolean. If data is from mouse, use TRUE, else FALSE.}
}
\value{
Character vector of gene symbols.
}
\description{
Convert excel dates to gene symbols
}
\examples{
x <- c("1-Dec","MYC","TP53")
y <- c("1-Dec","Myc","Tp53")
date_to_gene(x)
date_to_gene(y, mouse = TRUE)
}
