% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_get_started.R
\name{detect_species}
\alias{detect_species}
\title{Detect human or mouse data}
\usage{
detect_species(genes)
}
\arguments{
\item{genes}{Character vector of gene symbols.}
}
\value{
String stating the species detected.
}
\description{
Detect human or mouse data
}
\examples{
x <- c("MYC","BRAF","TP53")
y <- c("Myc","Braf","Tp53")
detect_species(x)
detect_species(y)
}
