% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helloworld.R
\name{helloBerlin}
\alias{helloBerlin}
\title{Say Hello}
\usage{
helloBerlin(message)
}
\arguments{
\item{message}{What do you want to say to Berlin?}
}
\description{
Say Hello, but only to Berlin
}
\details{
Sucks if you are not Berlin.
}
\examples{
helloBerlin("How are you?")
}
