% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_marker_analysis.R
\name{jaccard}
\alias{jaccard}
\title{Perform Jaccard equation on two vectors}
\usage{
jaccard(a, b)
}
\arguments{
\item{a}{Vector one.}

\item{b}{Vector two.}
}
\value{
Numeric vector.
}
\description{
Perform Jaccard equation on two vectors
}
