% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_sctype.R
\name{ScType_genesets}
\alias{ScType_genesets}
\title{List available ScType tissue gene sets}
\usage{
ScType_genesets()
}
\value{
Printed text.
}
\description{
List available ScType tissue gene sets
}
\examples{
ScType_genesets()
}
