% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_anno_summary.R
\name{berlin_anno_summary}
\alias{berlin_anno_summary}
\title{Write out BERLIN Annotation Summary}
\usage{
berlin_anno_summary(
  object = NULL,
  meta = NULL,
  file = NULL,
  excel = TRUE,
  return_seurat = FALSE,
  write_only = FALSE,
  anno_cols = c("^scType_.*\\\\_Classification$", "scType Score Sum", "ncells",
    "scType_Cell_Classification", "^scGate_.*\\\\_Celltype$", "_Jaccard_Classification$",
    "seurat_clusters", "RNA_snn_res.0.5", "RNA_snn_res.1", "RNA_snn_res.1.5",
    "RNA_snn_res.2", "dice.main", "dice.fine", "monaco.main", "monaco.fine",
    "northern.main", "northern.fine", "blue.main", "blue.fine")
)
}
\arguments{
\item{object}{Seurat object. If NULL, must supply meta data.}

\item{meta}{A data.frame. The rows are barcode/cell names, and the column are any additionally supplied metadata. Row names in the metadata need to match the column names of the counts matrix.}

\item{file}{String. Desired file name with directory for output.}

\item{excel}{Boolean. If TRUE, file will be save as excel workbook file.}

\item{return_seurat}{Boolean. If TRUE, a Seurat object will be the output with the results within the stored in the Misc slot.}

\item{write_only}{Boolean. If TRUE, function will only write out table to file, not to environment.}

\item{anno_cols}{Character Vector. Vector of column names to be subset and written out to file. List of default columns in details.}
}
\value{
data frame
}
\description{
Write out BERLIN Annotation Summary
}
\details{
Default annotation columns: c("^scType_.\emph{\\\emph{Classification$","scType Score Sum","ncells","scType_Cell_Classification",
"^scGate}.}\\_Celltype$","_Jaccard_Classification$","seurat_clusters","RNA_snn_res.0.5","RNA_snn_res.1",
"RNA_snn_res.1.5","RNA_snn_res.2","dice.main","dice.fine", "monaco.main","monaco.fine",
"northern.main","northern.fine","blue.main","blue.fine")
}
