% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_cluster_deg.R
\name{berlin_cluster_deg}
\alias{berlin_cluster_deg}
\title{Perform differential gene expression on clusters in a Seurat object}
\usage{
berlin_cluster_deg(
  object = NULL,
  assay = "RNA",
  cluster_cols = "seurat_clusters",
  verbose = TRUE,
  min.pct = 0.25,
  min.diff.pct = 0,
  adjpval_cutoff = 0.05,
  log2fc_cutoff = 1
)
}
\arguments{
\item{object}{Seurat object. If NULL, must supply count data.}

\item{assay}{String. Name of the initial/default assay. Default is RNA.}

\item{cluster_cols}{String or Vector. Column name from meta data of primary identified clusters. Default to 'seurat_clusters'. Multiple columns allowed.}

\item{verbose}{Boolean. To show progress, TRUE, else FALSE.}

\item{min.pct}{Numeric. Only test genes that are detected in a minimum fraction of cells. Default is 0.25.}

\item{adjpval_cutoff}{Numeric. An adjusted p-value value cutoff for identifying significantly expressed genes. Default is 0.05.}

\item{log2fc_cutoff}{Numeric. An absolute value of a log2 fold change cutoff to identify up and down regulated genes. Default is 0.}
}
\value{
A list of data frames containing DGE of clusters and a geneset of the markers of each cluster.
}
\description{
Perform differential gene expression on clusters in a Seurat object
}
