% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_cluster_deg.R
\name{berlin_dotplot}
\alias{berlin_dotplot}
\title{Generate dotplot heatmap of geneset gene expression within clusters}
\usage{
berlin_dotplot(
  object = NULL,
  markers_res = NULL,
  geneset = NULL,
  cluster_col = "seurat_clusters",
  assay = "RNA",
  min.pct = 0.25,
  scaled = FALSE,
  exp_binary = TRUE,
  log_exp = TRUE,
  adjpval_cutoff = 0.05,
  log2fc_cutoff = 1,
  verbose = TRUE,
  return_seurat = FALSE,
  plot_title = NULL,
  title_size = 20,
  x_text_size = 14,
  y_text_size = 14,
  x_title_size = 16,
  strip_text_size = 16,
  leg_title_size = 16,
  leg_text_size = 14
)
}
\arguments{
\item{object}{Seurat object. If NULL, must supply count data.}

\item{markers_res}{Data.frame. Result data from Seurat's FindAllMarkers() function.}

\item{geneset}{Data.frame or list. Two column data frame with first column consisting of geneset names and second column of the genes withing that geneset. One gene per row, the geneset name can repeat in the first column. If list input, this would be a named list where the names are the geneset names and each element a vector of genes as strings.}

\item{cluster_col}{String. Column name from meta data of primary identified clusters. Default to 'seurat_clusters'.}

\item{assay}{String. Name of the initial/default assay. Default is RNA.}

\item{min.pct}{Numeric. Only test genes that are detected in a minimum fraction of cells. Default is 0.25.}

\item{scaled}{Boolean. TRUE if plotting scaled expression data, FALSE if un-scaled.}

\item{exp_binary}{Boolean. If TRUE, size of dot will be binary, large if significantly upregulated, small if not. If FALSE, dot size will be based on percent of cells expressing gene.}

\item{log_exp}{Boolean. Only relevant if 'scaled' argument equals FALSE. If TRUE, average expression value will be logged with log2.}

\item{adjpval_cutoff}{Numeric. An adjusted p-value value cutoff for identifying significantly expressed genes. Default is 0.05.}

\item{log2fc_cutoff}{Numeric. An absolute value of a log2 fold change cutoff to identify up and down regulated genes. Default is 0.}

\item{verbose}{Boolean. To show progress, TRUE, else FALSE.}

\item{return_seurat}{Boolean. If TRUE, a Seurat object will be the output with the results within the stored in the Misc slot.}

\item{plot_title}{String. Desired plot title.}

\item{title_size}{Numeric. Title font size.}

\item{x_text_size}{Numeric. x-axis tick font size.}

\item{y_text_size}{Numeric. y-axis tick font size.}

\item{x_title_size}{Numeric. x-axis title font size.}

\item{strip_text_size}{Numeric. y-axis geneset annotation font size.}

\item{leg_title_size}{Numeric. Legend title font size.}

\item{leg_text_size}{Numeric. Legend text font size.}
}
\value{
ggplot object
}
\description{
Generate dotplot heatmap of geneset gene expression within clusters
}
