% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_qc.R
\name{berlin_qc}
\alias{berlin_qc}
\title{Perform quality control and initial formatting of Seurat object}
\usage{
berlin_qc(
  object = NULL,
  counts = NULL,
  meta = NULL,
  assay = "RNA",
  project_name = "BERLIN_Project",
  species = "human",
  seed = 42,
  percent_mt = NULL,
  minFeatures = 500,
  varFeatures = 2000,
  verbose = TRUE,
  logScale = 10000,
  remove_duplicates = TRUE,
  resolution = c(0.5, 1, 1.5, 2),
  pca_npcs = 30,
  tsne_dims = 1:30,
  umap_dims = 1:10,
  neighbor_dims = 1:30,
  doublet_pN = 0.25,
  doublet_pK = 0.09,
  doublet_prop = 0.04,
  doublet_PCs = 1:10
)
}
\arguments{
\item{object}{Seurat object. If NULL, must supply count data.}

\item{counts}{A matrix, data.frame, or dgCMatrix. The rows are unique features. The feature name can be the rowname or the first column. Each column is a uniquely labeled barcode/cell.}

\item{meta}{A data.frame. The rows are barcode/cell names, and the column are any additionally supplied metadata. Row names in the metadata need to match the column names of the counts matrix.}

\item{assay}{String. Name of the initial/default assay.}

\item{project_name}{String. Desired project name.}

\item{species}{String. Denoting if the data is from "human" or "mouse".}

\item{seed}{Numeric. Random seed, default is 42. Setting to NULL will remove seed.}

\item{percent_mt}{Numeric. Percent mitochondria to filter out. Default is NULL.}

\item{minFeatures}{Numeric. Minimum number of features a barcode/cell must have. Default to 500.}

\item{varFeatures}{Numeric. Number of features to select as top variable features. Default to 2000.}

\item{verbose}{Boolean. To show progress, TRUE, else FALSE.}

\item{logScale}{Numeric. Set the scale to log normalize data. Default to 10000.}

\item{remove_duplicates}{Boolean. To remove duplicate features that may be found in the data, use TRUE, else FALSE and the function will notify you if duplicate features are found.}

\item{resolution}{Numeric vector. Above one to obtain a larger number of communities, and below one to obtain a smaller number.}

\item{pca_npcs}{Numeric. Total number of principal components to compute for PCA plot. Default is 30.}

\item{tsne_dims}{Numeric sequence. Dimensions to use as input features for TSNE plot. Default 1:30.}

\item{umap_dims}{Numeric sequence. Dimensions to use as input features for UMAP. Default 1:10.}

\item{neighbor_dims}{Numeric sequence. Dimension of reduction to use as input to find neighbors.}

\item{doublet_pN}{Numeric. The number of generated artificial doublets, expressed as a proportion of the merged real-artificial data. Default is set to 0.25.}

\item{doublet_pK}{Numeric. The principal component neighborhood size to compute pANN, expressed as a proportion of the merged real-artificial data. Default is 0.09.}

\item{doublet_prop}{Numeric. Estimated proportion of homotypic doublets. Default is 0.04.}

\item{doublet_PCs}{Numeric. Number of statistically-significant principal components.}
}
\value{
Seurat object.
}
\description{
Perform quality control and initial formatting of Seurat object
}
