% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_scgate.R
\name{berlin_scgate_summarize}
\alias{berlin_scgate_summarize}
\title{Summarize scGate annotation across clusters}
\usage{
berlin_scgate_summarize(
  object = NULL,
  celltype_col = NULL,
  cluster_col = "seurat_clusters"
)
}
\arguments{
\item{object}{Seurat object. If NULL, must supply count data.}

\item{celltype_col}{String. Name of column containing predicted cell type names. If not supplied it will be predicted based on string matching.}

\item{cluster_col}{String. Name of column containing cluster information.}
}
\value{
A data.frame object.
}
\description{
Summarize scGate annotation across clusters
}
