% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_sctype.R
\name{berlin_sctype_classify}
\alias{berlin_sctype_classify}
\title{Assign cluster to scType}
\usage{
berlin_sctype_classify(
  object = NULL,
  score = NULL,
  meta = NULL,
  cluster_col = "seurat_clusters",
  scType_col = "scType_classification",
  verbose = TRUE
)
}
\arguments{
\item{object}{Seurat object.}

\item{score}{Matrix. Results from berlin_scType function, of a matrix with rownames as scTypes and columns of each cell and their associated scType enrichment score.}

\item{meta}{data.frame. If no Seurat object available, please provide meta data with the cluster column and rownames matching score cell names.}

\item{cluster_col}{String. Column name of meta data column that annotated clusters. Defaults to "seurat_clusters".}

\item{scType_col}{String. Desired column name for column to store scType classification results. Defaults to "scType_classification".}

\item{verbose}{Boolean. To show progress, TRUE, else FALSE.}
}
\value{
Seurat object with appended meta data.
}
\description{
Assign cluster to scType
}
