% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUMAPapp.R
\name{runUMAPapp}
\alias{runUMAPapp}
\title{Run UMAP Shiny App}
\usage{
runUMAPapp(
  object = NULL,
  counts = NULL,
  meta = NULL,
  n_cells = 2000,
  assay = "RNA",
  save_data = TRUE,
  save_to = getwd(),
  project_name = "BERLIN_Project",
  umap1_col = "UMAP_1",
  umap2_col = "UMAP_2",
  anno1_col = "seurat_clusters",
  anno2_col = NULL,
  anno3_col = NULL,
  seed = 42,
  species = "human",
  remove_duplicates = TRUE,
  launch.browser = TRUE,
  species_detected = "human",
  verbose = TRUE
)
}
\arguments{
\item{launch.browser}{Boolean logic to launch application in browser. Default to TRUE.}
}
\value{
An R Shiny Application.
}
\description{
Run UMAP Shiny App
}
